requirements_mac.txt
flask
flask-cors
python-dotenv
fpdf
pdfkit